refactor: AssimpTranslator에서 인덱스 카운트 타입 명시적으로 캐스팅하도록 수정

+++ EngineCore/Source/Asset/Pipeline/Translators/AssimpTranslator.cpp
@@ -84,7 +84,7 @@ void ProcessMesh(
     // Section(Submesh) 설정
     gfx::MeshSection section;
     section.index_start = 0;
-    section.index_count = pipeline_node.indices.Len();
+    section.index_count = static_cast<uint32>(pipeline_node.indices.Len());
     section.material_index = mesh->mMaterialIndex;
 
     pipeline_node.sections.Push(section);
